Show interpolated dashed lines instead of 'wells'.
It is common to see "wells" (when the line goes abrutly down and then
back up to approx. the same level it was before): these are due
to the user not having answered yet at the time of scores retrieval.
This causes him to "lose" lot of positions, and then regain them
the day after when he answers two question.
This commit add a functionality that checks for these kind of situations
and plots an interpolated dashed line instead of the "correct" one.